requirements.txt
fastapi
uvicorn[standard]
pydantic
websockets
langchain-core
autogen
groq
python-dotenv
networkx      